+++ cart/infrastructure/inMemoryCustomerCartserviceAdapter.go
@@ -3,8 +3,9 @@ package infrastructure
 import (
 	"context"
 
+	"flamingo.me/flamingo/v3/core/auth"
+
 	"flamingo.me/flamingo-commerce/v3/cart/domain/cart"
-	"flamingo.me/flamingo/v3/core/oauth/domain"
 )
 
 type (
@@ -26,18 +27,18 @@ func (gcs *InMemoryCustomerCartService) Inject(
 }
 
 // GetCart gets a customer cart from the in memory customer cart service
-func (gcs *InMemoryCustomerCartService) GetCart(ctx context.Context, auth domain.Auth, cartID string) (*cart.Cart, error) {
+func (gcs *InMemoryCustomerCartService) GetCart(ctx context.Context, identity auth.Identity, cartID string) (*cart.Cart, error) {
 	cart, err := gcs.inMemoryBehaviour.GetCart(ctx, cartID)
 	return cart, err
 }
 
 // GetModifyBehaviour gets the cart order behaviour of the service
-func (gcs *InMemoryCustomerCartService) GetModifyBehaviour(context.Context, domain.Auth) (cart.ModifyBehaviour, error) {
+func (gcs *InMemoryCustomerCartService) GetModifyBehaviour(context.Context, auth.Identity) (cart.ModifyBehaviour, error) {
 	return gcs.inMemoryBehaviour, nil
 }
 
 // RestoreCart restores a previously used cart
-func (gcs *InMemoryCustomerCartService) RestoreCart(ctx context.Context, auth domain.Auth, cart cart.Cart) (*cart.Cart, error) {
+func (gcs *InMemoryCustomerCartService) RestoreCart(ctx context.Context, identity auth.Identity, cart cart.Cart) (*cart.Cart, error) {
 	customerCart := cart
 
 	err := gcs.inMemoryBehaviour.storeCart(&customerCart)