+++ customer/module.go
@@ -2,10 +2,12 @@ package customer
 
 import (
 	"flamingo.me/dingo"
+	"flamingo.me/flamingo/v3/core/auth"
+	flamingoGraphql "flamingo.me/graphql"
+
 	customerDomain "flamingo.me/flamingo-commerce/v3/customer/domain"
 	customerInfrastructure "flamingo.me/flamingo-commerce/v3/customer/infrastructure"
 	customerGraphql "flamingo.me/flamingo-commerce/v3/customer/interfaces/graphql"
-	flamingoGraphql "flamingo.me/graphql"
 )
 
 type (
@@ -27,8 +29,14 @@ func (m *Module) Inject(config *struct {
 // Configure module
 func (m *Module) Configure(injector *dingo.Injector) {
 	if m.useNilCustomerAdapter {
-		injector.Bind((*customerDomain.CustomerService)(nil)).To(customerInfrastructure.NilCustomerServiceAdapter{})
 		injector.Bind((*customerDomain.CustomerIdentityService)(nil)).To(customerInfrastructure.NilCustomerServiceAdapter{})
 	}
 	injector.BindMulti(new(flamingoGraphql.Service)).To(customerGraphql.Service{})
+}
+
+// Depends on other modules
+func (m *Module) Depends() []dingo.Module {
+	return []dingo.Module{
+		new(auth.WebModule),
+	}
 }